Adding HTTP Error btn on list app
